# Revise Host index on inet_addresses