--- a/db/src/main/resources/sql/er_diagram/full_er_diagram.html
+++ b/db/src/main/resources/sql/er_diagram/full_er_diagram.html
@@ -261,7 +261,7 @@ td.section {
     </tr> 
     <tr> 
      <td class="property_name">generated on</td> 
-     <td class="property_value">2022-03-07 19:59:39.894792</td> 
+     <td class="property_value">2022-03-08 17:12:35.997985</td> 
     </tr> 
     <tr>
      <td class="property_name">last flyway file</td>
@@ -284,7 +284,7 @@ td.section {
      generated on
     </text> 
     <text text-anchor="start" x="4755.52" y="-10.8" font-family="Helvetica,sans-Serif" font-size="14.00">
-     2022-03-07 19:59:39.894792
+     2022-03-08 17:12:35.997985
     </text> 
     <polygon fill="none" stroke="#888888" points="4668.02,-4 4668.02,-44 4933.02,-44 4933.02,-4 4668.02,-4" /> <!-- allocationtoken_a08ccbef --> 
     <g id="node1" class="node"> 
@@ -11318,7 +11318,7 @@ td.section {
     <tr> 
      <td class="spacer"></td> 
      <td class="minwidth">inet_addresses</td> 
-     <td class="minwidth">ascending</td> 
+     <td class="minwidth">unknown</td> 
     </tr> 
     <tr> 
      <td colspan="3"></td> 
@@ -11342,18 +11342,6 @@ td.section {
     <tr> 
      <td class="spacer"></td> 
      <td class="minwidth">repo_id</td> 
-     <td class="minwidth">ascending</td> 
-    </tr> 
-    <tr> 
-     <td colspan="3"></td> 
-    </tr> 
-    <tr> 
-     <td colspan="2" class="name">idxovmntef6l45tw2bsfl56tcugx</td> 
-     <td class="description right">[non-unique index]</td> 
-    </tr> 
-    <tr> 
-     <td class="spacer"></td> 
-     <td class="minwidth">deletion_time</td> 
      <td class="minwidth">ascending</td> 
     </tr> 
    </tbody>