… db/src/main/resources/sql/er_diagram/brief_er_diagram.html